fix: fix PHP 8.5 deprecations (#170)
* fix: fix PHP 8.5 deprecations, add missing namespace in tests


* docs: update changelog


---------

+++ src/InfluxDB2/FluxCsvParser.php
@@ -283,7 +283,7 @@ class FluxCsvParser
             if ($strVal == '-Inf') {
                 return -INF;
             }
-            return (double)$strVal;
+            return (float)$strVal;
         }
 
         if ('base64Binary' == $column->dataType) {
